Create movie_stream2.py
 code added a new feature to the movie recommendation app, which allows users to select a movie based on their desired emotion, year, and duration.

It added a new option called "Random Movie" in the menu for the user to select, and added the following code snippets to implement the feature:

A selectbox to choose the emotion of the movie, using the unique values of the 'emotion' column in the dataframe.
A selectbox to choose the year of the movie, where the user can select either "classic" or "modern" movies.

A selectbox to choose the duration of the movie, where the user can select either "short" or "long" movies.

Filter the dataframe based on the selected values and display the filtered dataframe.