Create movie_stream3.py
his code adds some new streamlit widgets and formatting compared to the first and second code:

It uses the sidebar.selectbox() widget to allow the user to select the type of emotion, year, and duration to filter the movie.
It uses the sidebar.button() widget to allow the user to generate a random movie that matches the chosen filters.
It uses the markdown() method with some HTML styling to format the title of the randomly selected movie in a visually appealing way.
It uses the warning() method to display a warning message if there is no match found for the chosen filters or invalid inputs.
It uses the set_page_config() method to set the page title, icon, and layout of the app.

This code adds a few new features compared to the first and second code:

It uses two different databases "model_movies_data.db" and "emotion_movies_data.db"
It uses a new function called "filter_movie" that allows the user to filter movies based on the emotion, year, and duration, it also allows the user to generate a random movie that matches the chosen filters.
It has a new option to generate random movie in the interface, which when clicked generates a random movie that matches the chosen filters.
It uses the streamlit's set_page_config method to set the page title, icon and layout of the app.
It has some minor changes in the get_recommendation function, it gets the index of the movie by matching the title instead of using the pd.series() function.